==> 제 7회 롯데멤버스 빅데이터 경진대회/Code/00.Handling_Outside_Data.ipynb <==
# Import

import pandas as pd
import re
import numpy as np

# Data Load

weather_df = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/기상청 데이터 - 기온,강수량,습도.csv', encoding = 'CP949')
atmoshphere_df =  pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/기상청 데이터-미세먼지.csv', encoding = 'CP949')

# Data Preprocess

## Weather df

# 날짜 맞추기
weather_df['일시'] = weather_df['일시'].apply(lambda x: re.sub(r'[^\w\s]', '', x))

# 결측치 보간
weather_df.loc[4659, '평균기온(°C)'] = np.round(weather_df.iloc[4658:4661, 3].sum() / 2, 2)
weather_df.loc[5584, '평균기온(°C)'] = np.round(weather_df.iloc[5583:5586, 3].sum() / 2, 2)
weather_df.loc[5711, '평균기온(°C)'] = np.round(weather_df.iloc[5710:5713, 3].sum() / 2, 2)

# 결측치 보간
weather_df.loc[1159, '평균 상대습도(%)'] = np.round(weather_df.iloc[1158:1161, 5].sum() / 2, 2)
weather_df.loc[4784, '평균 상대습도(%)'] = 59.0
weather_df.loc[4785, '평균 상대습도(%)'] = 59.0
weather_df.loc[4659, '평균 상대습도(%)'] = np.round(weather_df.iloc[4658:4661, 5].sum() / 2, 2)
weather_df.loc[5584, '평균 상대습도(%)'] = np.round(weather_df.iloc[5583:5586, 5].sum() / 2, 2)
weather_df.loc[5711, '평균 상대습도(%)'] = np.round(weather_df.iloc[5710:5713, 5].sum() / 2, 2)

# 강수량 결측치 처리
weather_df.fillna(0, inplace = True)

def change_name(x):
    if x == '강릉':
        return 'Z01'
    elif x == '서울':
        return 'Z17'
    elif x == '인천':
        return 'Z04'
    elif x == '수원':
        return 'Z10'
    elif x == '대전':
        return 'Z15'
    elif x == '안동':
        return 'Z12'
    elif x == '대구':
        return 'Z06'
    elif x == '전주':
        return 'Z14'
    elif x == '울산':
        return 'Z03'
    elif x == '광주':
        return 'Z13'
    elif x == '부산':
        return 'Z16'
    elif x == '제주':
        return 'Z07'
    elif x == '제천':
        return 'Z09'
    elif x == '천안':
        return 'Z08'
    elif x == '세종':
        return 'Z02'
    elif x == '영광군':
        return 'Z05'
    elif x == '밀양':
        return 'Z11'

weather_df['지점명'] = weather_df['지점명'].apply(change_name)

weather_df = weather_df.iloc[:, 1:].rename(columns={'지점명': 'region',
                                      '일시': 'date',
                                      '평균기온(°C)': 'temperature',
                                      '일강수량(mm)': 'rainfall',
                                      '평균 상대습도(%)': 'huminity'})

## Atmos df

atmoshphere_df['일시'] = atmoshphere_df['일시'].apply(lambda x: re.sub(r'[^\w\s]', '', x))

def change_reg(x):
    if x == '서울':
        return '중부'
    elif x == '울산':
        return '남부'
    elif x == '고산':
        return '제주'

atmoshphere_df['지점명'] = atmoshphere_df['지점명'].apply(change_reg)

atmoshphere_df = atmoshphere_df.iloc[:, 1:].rename(columns = {'지점명': 'region_at',
                                                             '일시': 'date',
                                                             '일 미세먼지 농도(㎍/㎥)': 'atmosphere'})

# Deployment

weather_df.to_csv('../3.로티로리_데이터 및 모델 세이브 파일/Weather_df.csv'  ,header=True )
atmoshphere_df.to_csv('../3.로티로리_데이터 및 모델 세이브 파일/Atmoshphere_df.csv', header=True)
