==> 제 7회 롯데멤버스 빅데이터 경진대회/Code/06.ML.ipynb <==
# Import

# Data Handling
import pandas as pd
import numpy as np
import datetime
from sklearn.preprocessing import PowerTransformer
from sklearn.preprocessing import MinMaxScaler,StandardScaler
import category_encoders as ce
from sklearn import preprocessing
from sklearn.decomposition import PCA


# Visualization
import matplotlib.pylab as plt
%matplotlib inline


# OS
import os
import time
import warnings; warnings.filterwarnings("ignore")
from tqdm import tqdm


# Modeling
from sklearn.model_selection import KFold
n_splits=5; seed = 42
kfold = KFold(n_splits=n_splits, shuffle=True, random_state=seed)
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import RidgeClassifier, Lasso, ElasticNet
from sklearn.linear_model import BayesianRidge
from sklearn.neural_network import MLPClassifier


# Hyperparameter Optimization
from sklearn.model_selection import RandomizedSearchCV


# Ensemble
from itertools import combinations
from sklearn.ensemble import VotingClassifier


# Evaluation
from sklearn.model_selection import cross_val_score
from scipy.stats.mstats import gmean
from sklearn.metrics import roc_auc_score
from sklearn.metrics import accuracy_score


# Saving
import joblib

# Data Load

train = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/features12_tr.csv').drop('cust',axis=1)
test = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/features3_te.csv').drop('cust',axis=1)

print(train.shape)
display(train)

print(test.shape)
display(test)

# Preprocessing

pre_df = pd.concat([train,test]).reset_index(drop=True)

obj_col = ['favorite_pd','favorite_zon','favorite_loc','most_expensive','most_buying_weekday',
           'pdde_most_use_chnl','pdde_cop_c_most_use','cop_u_cop_c_most_use','most_buy_cop_u']

train.drop(obj_col,axis=1,inplace=True)
test.drop(obj_col,axis=1,inplace=True)

oh_f = pd.get_dummies(pre_df[obj_col])

max_col = oh_f.shape[1] # 차원축소하려는 데이터의 column의 개수
pca = PCA(n_components=max_col, random_state=0).fit(oh_f)

cumsum = np.cumsum(pca.explained_variance_ratio_) # 분산의 설명량을 누적합
num_col = np.argmax(cumsum >= 0.95) + 1 # 분산의 설명량이 99%이상 되는 차원의 수

# 차원축소
pca = PCA(n_components = num_col, random_state=0).fit_transform(oh_f)
pca_f = pd.DataFrame(pca)

pca_f

cf_tr = pca_f.iloc[:len(train),:].reset_index(drop=True)
cf_te = pca_f.iloc[len(train):,:].reset_index(drop=True)

train = pd.concat([train,cf_tr],axis=1)
test= pd.concat([test,cf_te],axis=1)

train

test

label_encoder=preprocessing.LabelEncoder()
X_train = train.drop('lpay_use_status',axis=1)
train_target = label_encoder.fit_transform(train['lpay_use_status'])
X_test = test.drop('lpay_use_status',axis=1)
test_target = label_encoder.transform(test['lpay_use_status'])

# Model

#### 1. 단일 모델 기본 성능 확인

# 단일모델 정의
knn_clf = KNeighborsClassifier(n_jobs=-1)
extra_clf = ExtraTreesClassifier(random_state=seed, n_jobs=-1)
gbm_clf = GradientBoostingClassifier(random_state=seed)
xgb_clf = XGBClassifier(random_state=seed, n_jobs=-1)
lgb_clf = LGBMClassifier(random_state=seed, n_jobs=-1)
cat_clf = CatBoostClassifier(random_state=seed, verbose=False)

clfs = [ridge_clf, knn_clf, extra_clf, gbm_clf, xgb_clf, lgb_clf, cat_clf]


# Cross_val_score 함수정의
def get_model_cv_prediction(model, feature_data, y_target):
    roc_auc_scores = cross_val_score(model, feature_data, y_target, scoring='roc_auc_ovr', cv=kfold, n_jobs=-1)
    avg_roc_auc = np.mean(roc_auc_scores)
    model = model.fit(feature_data,y_target)
    pred = model.predict_proba(X_test.values)#[:,1]
    test_roc = roc_auc_score(test_target,pred,multi_class='ovr',average='weighted')
    print(f'{model.__class__.__name__} 모델의 평균 AUC: {avg_roc_auc:.3f} test AUC : {test_roc:.3f}')


# 단일 모델별 평균 성능 출력
for clf in clfs:
    get_model_cv_prediction(clf, X_train.values, train_target)

#### 2. 랜덤서치를 활용한 모델 튜닝 

clfs = [
    (
        KNeighborsClassifier(),              
        {'n_neighbors': [3,5,7,9,11],        
         'weights': ['uniform','distance']}
    ),
    (
        ExtraTreesClassifier(),
        {'n_estimators': [100, 150, 200, 250, 300],
         'max_depth': [10, 12, 15, 17, 20],
         'max_features': [0.8, 0.85, 0.9, 0.95],
         'min_samples_split': [1, 2, 3, 4, 5],
         'min_samples_leaf': [1, 2, 3, 4, 5]}
    ),
    (
        GradientBoostingClassifier(),
        {
         'n_estimators': [100,300,500,1000],
         'learning_rate': [0.01,0.03,0.05,0.1],
         'max_depth': [3,5,6],
         'min_samples_leaf' : [3,5,7,9,10],
         'min_samples_split' : [2,4,6,8,10],
         'subsample' : [0.8,0.9,0.95,1]
         }
    ),
    (
        XGBClassifier(),
        {
        'n_estimators' :[100,200,300,400,500], # kaggle competition에서 best = 1000 
        'learning_rate': [0.01,0.03,0.05,0.1],
         'max_depth': [3,5,6],
         'colsample_bytree' :[0,0.1,0.3,0.5,0.7,0.9,1],
         'min_child_weight' :[1,3,5,6],
         'subsample' :[0.8,0.9,0.95,1],
         'objective' : ['clf:squarederror']
        }
    ),
    (
        LGBMClassifier(),
        {'n_estimators': [300,500,700,1000,1100],
         'learning_rate': [0.01,0.03,0.05,0.1],
         'max_depth': [3,5,7,9,10],
         #'boosting' : ['gbrt','dart'],
         'colsample_bytree' : [0,0.1,0.3,0.5,0.7,0.9,1],
         'subsample' :[0.8,0.9,0.95,1],
         'num_leaves' :[30,31,33,35,39,40]
         #'feature_fraction' : [0.1,0.3,0.5,0.7,0.9]
        }
    ),
    (
        CatBoostClassifier(),
        {'learning_rate': [0.05, 0.1, 0.2, 1, 1.5],
         'depth': [3, 5, 7, 9, 10],
         'iterations' : [500, 700, 1000, 1200],
         'verbose':[False]}
    )
]


RS_tuned_clfs = []  # 튜닝된 모델을 저장
for clf, param_grid in clfs:
    start = time.time()
    rand_search = RandomizedSearchCV(clf, param_grid, n_iter=20, scoring='roc_auc_ovr', 
                                     cv=kfold, random_state=seed, n_jobs=-1)
    rand_search.fit(X_train.values, train_target)
    clf_name = clf.__class__.__name__
    roc_score = rand_search.best_score_
    pred = rand_search.best_estimator_.predict_proba(X_test)
    test_roc = roc_auc_score(test_target,pred,multi_class='ovr',average='weighted')
    print(f'{clf_name:30s} mean_roc: {roc_score:.3f}, takes {time.time() - start:.1f} secs, test_roc: {test_roc:.3f}')
    RS_tuned_clfs.append((clf_name, rand_search.best_estimator_, roc_score,test_roc ))

# 튜닝된 모델들의 결과
RS_tuned_clfs

 - 튜닝된 모델들 저장

now = datetime.datetime.now()
nowDate = now.strftime('%Y-%m-%d')
joblib.dump(RS_tuned_clfs,f'../3.로티로리_데이터 및 모델 세이브 파일/RS_tuned_clfs_{nowDate}_pca_roc.pkl')

# 튜닝된 모델들 로드
RS_tuned_clfs = joblib.load('../3.로티로리_데이터 및 모델 세이브 파일/RS_tuned_clfs_2022-08-12_pca_roc.pkl')

#### 3. Catboost pred

pred = pd.DataFrame(RS_tuned_clfs[5][1].predict_proba(X_test))

pred.to_csv('../3.로티로리_데이터 및 모델 세이브 파일/Catboost_pred_df.csv',index=False)

#### 4. 튜닝된 모델들로 Voting 앙상블

# 모델 훈련 함수 정의
def return_fitted_model(model, train, target):
    acc_scores = []
    for iter_count, (train_idx, valid_idx) in enumerate(kfold.split(train, target)):

        X_train, X_valid = train[train_idx], train[valid_idx]
        y_train, y_valid = target[train_idx], target[valid_idx]

        model.fit(X_train, y_train)

        pred = model.predict(X_valid)
        acc_score = accuracy_score(y_valid, pred)
        acc_scores.append(acc_score)
    return model, np.mean(acc_scores)

# ex: extra_reg, extra_reg_score = return_fitted_model(extra_reg, train, target)
# print(f'모델의 평균 성능:  {extra_reg_score:.3f}')

 - 튜닝된 모델들 중 성능이 비슷하고 좋은 4개의 모델에 대해서 앙상블을 진행함

selected_clf = [
    #'RidgeClassifier',
    #'MLPClassifier',
    #'KNeighborsRegressor', 
    'ExtraTreesClassifier',
    #'CatBoostClassifier', 
    #'GradientBoostingClassifier', 
    'XGBClassifier',
    'LGBMClassifier'
]
models_for_ensemble = [(clf[0], clf[1]) for clf in RS_tuned_clfs if clf[0] in selected_clf]

 - 모든 조합에 대해서 앙상블을 수행

# 3개의 모델 중 2개씩 averaging, 3개씩 averging 모두 해보기
start = time.time()
best_avg_score = np.inf
bams = []
for model_nums in range(2, len(models_for_ensemble ) + 1):
    for avg_estimator in (combinations(models_for_ensemble , model_nums)):
        avg_reg = VotingClassifier(estimators = avg_estimator, n_jobs=-1,voting='soft')
        avg_model, avg_score = return_fitted_model(avg_reg, X_train.values, train_target)
        print(f'{"●".join([clf_name for clf_name, _, in avg_estimator])}: {avg_score:.3f}')
        if avg_score >= best_avg_score:
            best_avg_score = avg_score
            best_avg_reg = avg_model
        bams.append(avg_model)
print(time.time() - start)

 - 앙상블 모델 저장

now = datetime.datetime.now()
nowDate = now.strftime('%Y-%m-%d')
joblib.dump(bams[1],f'../3.로티로리_데이터 및 모델 세이브 파일/Average_Ensemble_{nowDate}_voting.pkl')

# 앙상블 모델 로드
best_avg_reg = joblib.load('../3.로티로리_데이터 및 모델 세이브 파일/Average_Ensemble_2022-08-12_voting.pkl')

- model 성능

pred = pd.DataFrame(bams[1].predict_proba(X_test))
print('Voting Ensemble roc_auc',roc_auc_score(test_target,pred,multi_class='ovr',average='weighted'))

#### 5. 최종 서브미션 저장

pred.to_csv('../3.로티로리_데이터 및 모델 세이브 파일/ML_Voting_pred_df.csv',index=False)
