==> 제 7회 롯데멤버스 빅데이터 경진대회/Code/07.Ensemble.ipynb <==
# Import

import pandas as pd
from sklearn.metrics import roc_auc_score

# Data Load

test = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/features3_te.csv')
tabnet = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/TabNet_pred_df.csv').drop('cust',axis=1)
pycaret = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/Pycaret_pred_df.csv').drop('cust',axis=1)
Catboost = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/Catboost_pred_df.csv')
ML = pd.read_csv('../3.로티로리_데이터 및 모델 세이브 파일/ML_Voting_pred_df.csv')

test

tabnet.columns = pycaret.columns
tabnet

pycaret

Catboost.columns = pycaret.columns
Catboost

ML.columns = pycaret.columns
ML

# Ensemble

ensemble_df = (tabnet+pycaret+Catboost+ML)/4
ensemble_df

test_target = test.lpay_use_status

print('Avg_ensemble roc_auc score :',roc_auc_score(test_target,ensemble_df,multi_class='ovr',average='weighted'))

# Final Pred

ensemble_df.to_csv('../3.로티로리_데이터 및 모델 세이브 파일/Final_pred.csv',index=False)
